+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Upcoming
+
+* [FIX] event matching fails if any event props pattern is not specified
+
 ## v1.0.0-alpha.6 - 2017-06-13
 
 * [FEATURE] allow interaction with the DeskPro Window via Deskpro Window API